requirements.txt
streamlit==1.31.0
scikit-learn==1.4.0
opencv-python==4.9.0.80
Pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
supabase==2.3.4
python-dotenv==1.0.1
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.2